pyproject.toml
[project]
name = "myproject"
version = "0.1.0"
description = "Laundry Platform"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "Django==6.0.2",
    "django-allauth==65.14.3",
    "django-crispy-forms==2.6",
    "crispy-bootstrap5==2026.3",
    "django-htmx==1.27.0",
    "django-environ==0.13.0",
    "dj-database-url==3.1.2",
    "psycopg2-binary==2.9.11",
    "whitenoise==6.12.0",
    "python-dotenv==1.2.2",
    "django-otp==1.6.1",
    "django-timezone-field==7.2.1",
    "django-phonenumber-field[phonenumbers]==8.4.0",
    "django-tinymce==5.0.0",
    "djangorestframework==3.16.1",
    "django-cors-headers==4.9.0",
    "django-redis==6.0.0",
    "attrs>=26.1.0",
    "requests>=2.31.0",
    "Pillow>=10.0.0",
    "gunicorn>=21.0.0",
    "PyJWT>=2.8.0",
    "cryptography>=41.0.0",
    "dj-database-url>=3.1.2",
    "chromadb>=0.4.0",
    "qrcode>=7.3.1",
    "googlemaps<=4.10.0"
]
